Reworked planck_2020_CamSpec_python to be inherited from the 2018 version, and added all necessary default yaml files. Possibly still buggy.
Made a new directory, likelihoods/planck_2020_highl_CamSpec, to contain the new class.
All changes from _base_classes/_planck_2020_CamSpec_python.py moved to planck_2020_CamSpec_python.py in the new directory
Fixed bug where amplitudes were not being used in power law residuals

yaml defaults based on ini files from CosmoMC. Used CamSpec2018 yaml files as a template.

Data files should be placed in a directory CamSpec2020 in the default cobaya data directory
foreground files are still required, even if powerlaw residuals are used instead.
