+++ cobaya/likelihoods/planck_2020_highl_CamSpec/TTTEEE_CS2020.py
@@ -1,0 +1,4 @@
+from cobaya.likelihoods.planck_2020_highl_CamSpec.planck_2020_CamSpec_python import planck_2020_CamSpec_python
+
+class TTTEEE_CS2020(planck_2020_CamSpec_python):
+    pass
