+++ cobaya/likelihoods/planck_2020_highl_CamSpec/TT_CS2020.bibtex
@@ -1,0 +1,16 @@
+high-$\ell$ temperature-only \textsc{CamSpec} likelihood of Planck's 2018 data release \cite{efstathiougratton} (native Python re-implemetation, by A.~Lewis)
+@ARTICLE{efstathiougratton,
+       author = {{Efstathiou}, George and {Gratton}, Steven},
+        title = "{A Detailed Description of the CamSpec Likelihood Pipeline and a Reanalysis of the Planck High Frequency Maps}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, High Energy Physics - Theory},
+         year = 2019,
+        month = oct,
+          eid = {arXiv:1910.00483},
+        pages = {arXiv:1910.00483},
+archivePrefix = {arXiv},
+       eprint = {1910.00483},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv191000483E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
