+++ cobaya/likelihoods/planck_2020_highl_CamSpec/TT_CS2020.py
@@ -1,0 +1,4 @@
+from cobaya.likelihoods.planck_2020_highl_CamSpec.planck_2020_CamSpec_python import planck_2020_CamSpec_python
+
+class TT_CS2020(planck_2020_CamSpec_python):
+    pass
