+++ cobaya/likelihoods/planck_2020_highl_CamSpec/params_EE_CamSpec.yaml
@@ -1,0 +1,11 @@
+  calEE:
+    prior:
+      dist: norm
+      loc: 1
+      scale: 0.02
+    ref:
+      dist: norm
+      loc: 1
+      scale: 0.005
+    proposal: 0.005
+    latex: c_{EE}
