+++ cobaya/likelihoods/planck_2020_highl_CamSpec/params_TT_CamSpec_fixedcalpol.yaml
@@ -1,0 +1,2 @@
+  calEE: 1
+  calTE: 1
