+++ cobaya/likelihoods/planck_2020_highl_CamSpec/params_calib_CamSpec_2020.yaml
@@ -1,0 +1,12 @@
+  A_planck:
+    prior:
+      dist: norm
+      loc: 1
+      scale: 0.0025
+    ref:
+      dist: norm
+      loc: 1
+      scale: 0.002
+    proposal: 0.0005
+    latex: y_\mathrm{cal}
+    renames: calPlanck
